+++ src/index.html
@@ -1,19 +1,19 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-	<meta charset="UTF-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<link rel="preconnect" href="https://fonts.googleapis.com">
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="preconnect" href="https://fonts.googleapis.com">
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap" rel="stylesheet">
-	<link rel="stylesheet" href="./style.css">
-	<title>Virtual Keyboard</title>
+  <link rel="stylesheet" href="./style.css">
+  <title>Virtual Keyboard</title>
 </head>
 <body>
-	
+  <script type="module" src="./scripts/layout.js"></script>
+  <script type="module" src="./scripts/script.js"></script>
+  <script type="module" src="./scripts/virtual.js"></script>
+  <script type="module" src="./scripts/handlers.js"></script>
 </body>
-<script type="module" src="./script.js"></script>
-<script type="module" src="./virtual.js"></script>
-<script type="module" src="./scripts/handlers.js"></script>
 </html>
